some more not working tests
