+++ lib/extensions/queue-me.js
@@ -1,4 +1,5 @@
 var request = require("request"),
+    serviceHelper = require("./service-helper.js");
     extensionHelper = require("./extension-helper.js");
 
 var queueMeExtension = function(keyword, playQueue) {
@@ -11,7 +12,7 @@ var queueMeExtension = function(keyword, playQueue) {
                 return;
             }
 
-            var service = getService(param);
+            var service = serviceHelper.getService(param);
             var videoId = param.split(".")[2].split("=")[1];
             var personName = message.data.name;
             var room = message.channel.slice(message.channel.lastIndexOf("/") + 1);
@@ -44,10 +45,6 @@ var queueMeExtension = function(keyword, playQueue) {
         }
     };
 };
-
-function getService(url) {
-    return url;
-}
 
 function convert_time(duration) {
     var a = duration.match(/\d+/g);
