+++ lib/extensions/service-helper.js
@@ -1,0 +1,5 @@
+function getService(url) {
+return url;    
+};
+
+module.exports = getService;
